safari-600.7-branch/LayoutTests/accessibility/aria-flowto-expected.txt

Item 1
Item 2
Item 3
BUTTON
RADIO BUTTON
This tests that aria-flowto correctly identifies the right elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item.ariaFlowToElementAtIndex(0).role is 'AXRole: AXButton'
PASS item.ariaFlowToElementAtIndex(0).title is 'AXTitle: BUTTON'
PASS item.ariaFlowToElementAtIndex(1).role is 'AXRole: AXRadioButton'
PASS item.ariaFlowToElementAtIndex(1).title is 'AXTitle: RADIO BUTTON'
PASS successfullyParsed is true

TEST COMPLETE
